Store money & quantity as fixed-point INTEGER (*1e8)
Replace float (REAL) money/quantity columns with scaled 64-bit integers
(_e8 = value * 1e8), the SQLite equivalent of PostgreSQL NUMERIC(38,8).
This eliminates float rounding drift in balances and P&L.

- Add db/money.py: Decimal<->int boundary (to_e8/from_e8/q/dec), prec=38,
  ROUND_HALF_UP; round once at the DB boundary.
- schema.sql: ledger columns -> *_e8 with CHECKs; keep market-data caches
  (price_snapshots, ohlcv_cache) as REAL. Also: move funnel_cycles before
  its referencing tables, add UNIQUE(user_id) on accounts, add indexes on
  transactions(user_id, executed_at), news(ticker, published_at),
  leaderboard(user_id, snapshot_at), corporate_actions(ticker, effective_date).
- Models expose Decimal; execution engine does all trade math in Decimal so
  cash_after = cash_before -/+ total_value always balances.
- server.py: Decimal-aware JSON encoder for responses and WebSocket.

Validation: tests/ 23 passed; integrity_check 36/0 (incl. cash reconciliation).

+++ config.py
@@ -4,6 +4,7 @@ All tunable parameters live here. Sensitive values loaded from .env.
 """
 
 import os
+from decimal import Decimal
 from pathlib import Path
 from dotenv import load_dotenv
 
@@ -34,7 +35,7 @@ DB_PATH = PROJECT_ROOT / "data" / "portfolio.db"
 SCHEMA_PATH = PROJECT_ROOT / "db" / "schema.sql"
 
 # ── Simulation Parameters ────────────────────────────────
-STARTING_BALANCE = float(os.getenv("STARTING_BALANCE", "10000.00"))
+STARTING_BALANCE = Decimal(os.getenv("STARTING_BALANCE", "10000.00"))
 STARTING_CURRENCY = "USD"
 
 # ── Funnel Configuration ─────────────────────────────────
